requirements.txt
beautifulsoup4==4.7.1
certifi==2019.3.9
chardet==3.0.4
colorful==0.5.0
Faker==1.0.5
idna==2.8
lxml==4.3.3
Pillow==6.0.0
prettyprinter==0.17.0
PyExecJS==1.5.1
Pygments==2.3.1
python-dateutil==2.8.0
requests==2.21.0
selenium==3.141.0
six==1.12.0
soupsieve==1.9.1
text-unidecode==1.2
urllib3==1.24.2
xmltodict==0.12.0
